29.107 s -> jestChangedFiles.*.ts (13.992 s) jestChangedFilesGit.*.ts (16.094 s)
